imageRepository/imageRepository/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-money==1.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py-moneyed==0.8.0
PyJWT==1.7.1
pymongo==3.11.1
pytz==2020.4
six==1.15.0
sqlparse==0.2.4
whitenoise==5.2.0
